integrated route for admin

+++ OnlineRailwayReservation/src/main/java/com/sunbeam/dao/RouteDao.java
@@ -8,8 +8,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 
 public interface RouteDao extends JpaRepository<Route, Integer>{
-	Route findById(int Id);
+	Route findById(Long Id);
 	List<Route>findAll();
 	Route save(Route route);
-	boolean deleteById(int Id);
+	boolean deleteById(Long Id);
 }
